[Utils cleanup] Rename einsum utility file to einsum
