+++ backpack/core/derivatives/avgpool2d.py
@@ -5,7 +5,7 @@ import torch.nn
 from torch.nn import AvgPool2d, Conv2d, ConvTranspose2d
 
 from ...utils import conv as convUtils
-from ...utils.utils import einsum
+from ...utils.einsum import einsum
 from .basederivatives import BaseDerivatives
 from backpack.utils.unsqueeze import jmp_unsqueeze_if_missing_dim
