+++ backpack/core/derivatives/batchnorm1d.py
@@ -1,6 +1,6 @@
 from torch.nn import BatchNorm1d
 
-from ...utils.utils import einsum
+from ...utils.einsum import einsum
 from .basederivatives import BaseDerivatives
 from backpack.utils.unsqueeze import jmp_unsqueeze_if_missing_dim
