+++ backpack/core/derivatives/conv2d.py
@@ -4,7 +4,7 @@ from torch.nn.functional import conv2d, conv_transpose2d
 
 from ...core.layers import Conv2dConcat
 from ...utils import conv as convUtils
-from ...utils.utils import einsum
+from ...utils.einsum import einsum
 from .basederivatives import BaseDerivatives
 from backpack.utils.unsqueeze import jmp_unsqueeze_if_missing_dim
