+++ backpack/core/derivatives/elementwise.py
@@ -1,4 +1,4 @@
-from ...utils.utils import einsum
+from ...utils.einsum import einsum
 from .basederivatives import BaseDerivatives
 from backpack.utils.unsqueeze import jmp_unsqueeze_if_missing_dim
