+++ backpack/core/derivatives/zeropad2d.py
@@ -1,7 +1,7 @@
 from torch.nn import ZeroPad2d
 from torch.nn.functional import pad
 
-from ...utils.utils import einsum
+from ...utils.einsum import einsum
 from .basederivatives import BaseDerivatives
 from backpack.utils.unsqueeze import jmp_unsqueeze_if_missing_dim
