+++ backpack/extensions/firstorder/batch_l2_grad/linear.py
@@ -1,4 +1,4 @@
-from backpack.utils.utils import einsum
+from backpack.utils.einsum import einsum
 from backpack.extensions.firstorder.base import FirstOrderModuleExtension
