+++ backpack/extensions/secondorder/diag_ggn/conv2d.py
@@ -1,5 +1,5 @@
 from backpack.utils import conv as convUtils
-from backpack.utils.utils import einsum
+from backpack.utils.einsum import einsum
 from backpack.core.derivatives.conv2d import Conv2DDerivatives, Conv2DConcatDerivatives
 from .diag_ggn_base import DiagGGNBaseModule
