+++ backpack/extensions/secondorder/diag_ggn/linear.py
@@ -1,5 +1,5 @@
 from backpack.core.derivatives.linear import LinearDerivatives, LinearConcatDerivatives
-from backpack.utils.utils import einsum
+from backpack.utils.einsum import einsum
 from .diag_ggn_base import DiagGGNBaseModule
