+++ backpack/extensions/secondorder/hbp/linear.py
@@ -1,5 +1,5 @@
 from backpack.core.derivatives.linear import LinearDerivatives, LinearConcatDerivatives
-from backpack.utils.utils import einsum
+from backpack.utils.einsum import einsum
 from .hbpbase import HBPBaseModule
 from .hbp_options import BackpropStrategy, ExpectationApproximation
